added single factory test

+++ src/domain/condiments/Condiment.java
@@ -2,9 +2,9 @@ package domain.condiments;
 
 public class Condiment {
 	public String name;
-	public int qty;
+	public long qty;
 	
-	public Condiment(String name, int qty) {
+	public Condiment(String name, long qty) {
 		this.name = name;
 		this.qty = qty;
 	}
